backend/requirements.txt
backend==0.2.4.1
chromadb==1.4.0
fastapi==0.128.0
langchain_community==0.4.1
langchain_core==1.2.6
langchain_huggingface==1.2.0
langchain_ollama==1.0.1
langchain_text_splitters==1.1.0
numpy==2.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
python-dotenv==1.2.1
python_jose==3.5.0
rank_bm25==0.2.2
sentence_transformers==5.2.0
SQLAlchemy==2.0.45
torch==2.9.1
pymysql
pydantic[email]
python-multipart
pypdf
